🟢 replicating bash test in typescript
Lots of redundancies
